requirements.txt

django>=2
djangorestframework
